:recycle: move to useReplayData
